+++ atcg_lib/include/Scene/ComponentGUIHandler.h
@@ -7,60 +7,69 @@
 
 namespace atcg
 {
+namespace GUI
+{
+
 /**
  * @brief A class that handles the rendering of gui components
  *
- * To add custom rendering code, create a class that inherits from this class and add the rendering code for the custom
- * component. To get the default behavior, override the draw_component function that calls the super class. Add this
- * class as template argument to the atcg::SceneHierarchyPanel.
- *
- * @note The code below is an example for MSVC. Under different compilers, one might have to restructure the template
- * specialization.
+ * To add custom rendering code, create a class that specializes this class and add the rendering code for the custom
+ * component.
  *
  * @code{.cpp}
- * class MyComponentGUIHandler : public ComponentGUIHandler
+ * template<>
+ * struct atcg::GUI::ComponentGUIRenderer<CustomComponent>
  * {
- * public:
- *      MyComponentGUIHandler(const atcg::ref_ptr<Scene>& scene) :ComponentGUIHandler(scene) {}
- *
- *      template<typename T>
- *      void draw_component(Entity entity, T& component)
- *      {
- *          atcg::ComponentGUIHandler::draw_component<T>(entity, component);
- *      }
- *
- *      template<>
- *      void draw_component<MyCustomComponent>(Entity entity, MyCustomComponent& component)
- *      {
- *          // Draw custom component
- *      }
+ *     void draw_component(const atcg::ref_ptr<Scene>& scene, Entity entity, CustomComponent& component) const
+ *     {
+ *         // Render Code
+ *     }
  * };
  * @endcode
  */
-class ComponentGUIHandler
+template<typename T>
+struct ComponentGUIRenderer
 {
-public:
     /**
-     * @brief Constructor
+     * @brief Draw the component
      *
-     * @param scene The scene
+     * @param scene The scene the entity belongs to
+     * @param entity The entity that holds the component
+     * @param component The component to render
      */
-    ComponentGUIHandler(const atcg::ref_ptr<Scene>& scene) : _scene(scene) {}
-
-    /**
-     * @brief Draw a component
-     *
-     * @tparam T The component type
-     *
-     * @param entity The entity that owns the component
-     * @param component The component
-     */
-    template<typename T>
-    void draw_component(Entity entity, T& component);
-
-protected:
-    bool displayMaterial(const std::string& key, Material& material);
-
-    atcg::ref_ptr<Scene> _scene;
+    void draw_component(const atcg::ref_ptr<Scene>& scene, Entity entity, T& component) const
+    {
+        throw std::logic_error("No ComponentGUIRenderer specialization available for this component type");
+    }
 };
+
+#define ATCG_DECLARE_COMPONENT_GUI_RENDERER(ComponentType)                                                             \
+    template<>                                                                                                         \
+    struct ComponentGUIRenderer<ComponentType>                                                                         \
+    {                                                                                                                  \
+        void draw_component(const atcg::ref_ptr<Scene>& scene, Entity entity, ComponentType& component) const;         \
+    }
+
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(TransformComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(CameraComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(GeometryComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(MeshRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(PointRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(PointSphereRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(EdgeRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(EdgeCylinderRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(InstanceRenderComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(PointLightComponent);
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(ScriptComponent);
+
+/**
+ * @brief Display a material
+ *
+ * @param key The imgui key
+ * @param material The material
+ *
+ * @return true if the material has been updated this frame
+ */
+bool displayMaterial(const std::string& key, Material& material);
+}    // namespace GUI
 }    // namespace atcg